StargazerServer-AlpacaClient/requirements.txt
skyfield
pytz
astropy
numpy
requests
pandas
pyserial
redis
# May need! python-dateutil